pyproject.toml
[project]
name = "ai-fake-image-detection"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "torch>=2.0",
    "torchvision>=0.15",
    "transformers>=4.36",
    "loralib>=0.1.2",
    "streamlit>=1.32",
    "opencv-python>=4.8",
    "pillow>=10.0",
    "pyyaml>=6.0",
    "scikit-image>=0.22",
    "timm>=0.9",
    "tqdm>=4.66",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "imutils>=0.5.4",
    "watchdog>=6.0.0",
    "efficientnet-pytorch>=0.7.1",
    "huggingface-hub>=1.11.0",
    "torchmetrics>=1.9.0",
    "wandb>=0.26.0",
    "datasets>=4.8.4",
    "torchaudio>=2.11.0",
    "openai>=1.30",
    "python-dotenv>=1.0",
]
